+++ CHANGELOG.md
@@ -41,6 +41,7 @@ Uliweb Change Log
   get_model to get those models.
 * Make get_model() set_model() don't distinquish case
 * Fix expose for two different view functions but with the same URL does not keep the last one bug
+* Fix rabc init not using functions.has_role and functions.has_permission.
 
 0.2.1 Version
 -----------------